Move market data fetching logic to server-side API route
